phosphor buttons: Enable Delayed/Fast Power Off (#1)
The current Power Button behavior is to power On or Off host with a
short press. The power Off behaviour on FSP-based platforms is
different, and is follows:
Initial 4 seconds: do nothing
Delayed Power Off Initiated: With this, 10 seconds countdown begins.
If Button is released during this interval and not pressed again,
normal shutdown - equivalent to 'obmcutil poweroff' - starts and
the host goes to Off state. If the Button is released and pressed
again during this interval OR button is held down beyond this interval,
Fast Power Off is initiated with 10 seconds countdown, and the host is
shutdown - equivalent of 'obmcutil chassisoff'.
This commit changes the behavior on p10bmc platforms similar to
FSP-based platforms.

The following tests were performed using simulator signals and
power button on rainier system:
1. Power On host (push/release button)
   Verified: host in 'Running' state
2. Power Off (Release Button within 4s)
   Verified: No change
3. Power Off (Release Button somewhere between 5-9s)
   Verified: host in 'Off' state
4. Power Off (Release Button after 6s, and press/release again)
   Verified: host in 'Off' state
5. Power Off (Release Button after 6s, press and hold it down 15s)
   Verified: host in 'Off' state
6. Power Off (Release after 30s)
   Verified: host in 'Off' state

Signed-off-by: Shantappa Teekappanavar <sbteeks@yahoo.com>
Change-Id: If8e4955ee748edb50df458dd64eb805b8406ea8a